move to standard repository structure
