+++ autobrewer/GUI/DeviceStatusGUI.py
