+++ autobrewer/GUI/UI Files/DeviceStatus.ui
