+++ autobrewer/GUI/examplePair/example.ui
