+++ autobrewer/MainWindow.py
@@ -50,8 +50,8 @@ class MainWindow(QtWidgets.QMainWindow):
 
         self.brewConfigScreen.StartBrewButton.clicked.connect(partial(self.goToMenu, self.brewStatusScreen))
         self.mainMenu.EnterBrewConfigButton.clicked.connect(partial(self.goToMenu, self.brewConfigScreen))
-        self.mainMenu.EnterCleanScreenButton.clicked.connect(partial(self.goToMenu, self.CleaningScreen))
-        self.mainMenu.EnterDeviceStatusScreen.clicked.connect(partial(self.goToMenu, self.DeviceStatusScreen))
+        self.mainMenu.EnterCleanScreenButton.clicked.connect(partial(self.goToMenu, self.cleaningScreen))
+        self.mainMenu.EnterDeviceStatusScreen.clicked.connect(partial(self.goToMenu, self.deviceStatusScreen))
         logger.debug("Menu navigation buttons connected")
